RSA: Unify interface to GMP ("_fastmath") and pure python ("_slowmath") implementations
This commit adds a new module called "_slowmath", which mirrors the RSA
portions of _fastmath.  A single, unified interface is layered on top of it.
This should help avoid any inconsistencies between the _fastmath and the
_slowmath code paths.

Also, users no longer need to specify an empty second parameter ("K") to
rsaObj.encrypt(), and specifying a non-empty second parameter results in a
DeprecationWarning.

Using the RSA primitive is now as simple as:

    from Crypto.PublicKey import RSA

    rsaObj = RSA.generate(2048)      # randomly generate a 2048-bit RSA key
    (ciphertext,) = rsaObj.encrypt("message")

Note that this is still just the RSA _primitive_.  PKCS#1 padding will come
later.