RSA: Make .verify() accept extra garbage after the signature.
The developers of Twisted Conch (an SSH protocol implementation) apparently
think they need to add extra junk to the end of the signature tuple when
calling RSAobj.verify.  In other words, they do something like this:

    RSAobj.verify("foo", (sig, ''))

instead of something like this:

    RSAobj.verify("foo", (sig,))

This isn't necessary, but it worked in PyCrypto 2.0.1.

The people behind Twisted Conch probably got confused by a similar requirement
in RSAobj.sign().  I could call it "user error" and rebuke them for misusing
the API, but that would be dumb.  The unified Crypto.PublicKey API is
confusing, encouraging exactly this kind of mistake.

The Crypto.PublicKey API needs to be replaced with something less error-prone,
but in the meantime, I am applying this change to make the .verify() method
behave how Twisted Conch expects.