Bugfix #1119552: PKCS#1v1.5 has to accept signatures without NULL parameters
The digest AlgorithmIdentifier has optional (NULL) parameters; the
verification function should not reject a signature if they are omitted.

With this fix, either case is acceptable (parameters present with value NULL
or not present).

As an exception, signatures based on old MD2/MD5 must always have NULL
parameters.

See Appendix B.1 of RFC 3447 and Section 2.1 of RFC 4055.

Closes: https://bugs.launchpad.net/pycrypto/+bug/1119552
[dlitz: Rebased and updated to use refactored asn1 API, text OIDs, & to fix Python 2.1.]