Counter: Deprecate disable_shortcut; Remove __PCT_CTR_SHORTCUT__ entirely
The `disable_shortcut` option served as a workaround in case
`__PCT_CTR_SHORTCUT__` leaked through a wrapper object, but I don't
think anyone actually used it, and it was a bad idea to expose it as
part of the public API.

Now that we do strong type checking inside block_template.c, there
shoujld be no need to ever use this option.  It's now a no-op, retained
for backward compatibility only.  It will be removed in some future
version of PyCrypto.