Add support for deterministic DSA.
This patch implements the variant of DSA
described in http://tools.ietf.org/html/draft-pornin-deterministic-dsa-02.

The nonce k is not taken from the RNG: instead, it is derived from
the message and the key.

DSA is still secure even on platforms where the RNG is not reliable
(e.g. in VMs).