Simplify _fastmath (remove import)
The randfunc parameter of getRandomInteger() can
be a callable object or None.

If it is None, C code will import the Random module
and create a new RNG object.

Apart from breaking pypy, the C code is ugly and unnecessary:
it is much easier to ensure that getRandomInteger()
is always called with a valid object from the Python side.