Workaround parallel build failures
Use unique source file names to fix a race condition due to a distutils
bug.  The distutils build system does not handle the same source files
being used by multiple extensions correctly.  If parallel (-j) builds
are being done, the same file is compiled twice to the same output file.
This creates a race condition: if the second compilation starts while
the first extension is being linked, it is possible for the linker to
grab the incomplete file that is being written by the second compiler
call.

The simplest workaround for this is to create additional files that
include the original sources while providing a unique name for them.

Fixes #378